Moved uploading attachments to the view model
